Update library use
Protobuf is no longer a submodule. Instead, it is automatically downloaded
by the setup scripts for windows, and automatically installed as a package
by the setup scripts for linux.

The Qt project file is now fixed to use libraries correctly.

The build and setup scripts are updated accordingly, and use PowerShell on
windows.

The README is updated to reflect these changes.